requirements.txt
alabaster==0.7.12
argcomplete==1.12.3
arrow==1.2.1
astroid==2.6.6
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
black==21.10b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cookiecutter==1.7.3
cryptography==35.0.0
deap==1.3.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
docutils==0.17.1
entrypoints==0.3
flake8==3.9.2
idna==3.3
imagesize==1.2.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.10.0
jedi==0.18.0
Jinja2==3.0.2
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.2.0
jupyter-client==6.1.12
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.2.1
lazy-object-proxy==1.6.0
lightgbm==3.3.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
numpy==1.21.3
numpydoc==1.1.0
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.8.0
parso==0.8.2
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.21
psutil==5.8.0
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
pylint==2.9.6
pyls-spyder==0.4.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
pytz==2021.3
pywin32==302
pywin32-ctypes==0.2.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.1.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.11.2
requests==2.26.0
rope==0.21.0
Rtree==0.9.7
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn-genetic==0.4.1
snowballstemmer==2.1.0
sortedcontainers==2.4.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.1.5
spyder-kernels==2.1.3
statsmodels==0.13.1
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.2
threadpoolctl==3.0.0
three-merge==0.1.1
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
yapf==0.31.0
zipp==3.6.0
